feat: access to detail page and some refactoring pn 845 (#35)
