+++ packages/pn-personafisica-webapp/src/navigation/routes.tsx
@@ -1,6 +1,7 @@
 import { Routes, Route } from 'react-router-dom';
 import { NotFound } from '@pagopa-pn/pn-commons';
 
+import NotificationDetail from '../pages/NotificationDetail.page';
 import Notifiche from '../pages/Notifiche.page';
 import RequireAuth from './RequiredAuth';
 import VerifyUser from './VerifyUser';
@@ -13,6 +14,7 @@ function Router() {
       {/* protected routes */}
       <Route element={<RequireAuth />}>
         <Route path={routes.NOTIFICHE} element={<Notifiche/>} />
+        <Route path={routes.DETTAGLIO_NOTIFICA} element={<NotificationDetail />} />
       </Route>
       <Route path="*" element={<NotFound />} />
     </Routes>
