+++ packages/pn-personafisica-webapp/src/redux/notification/reducers.ts
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
+import { NotificationStatus } from '@pagopa-pn/pn-commons';
 import {
-  NotificationStatus,
   NotificationDetail,
   NotificationDetailDocument,
   NotificationDetailPayment,
@@ -9,6 +9,7 @@ import {
   NotificationStatusHistory,
   PhysicalCommunicationType,
 } from '@pagopa-pn/pn-commons';
+
 import {
   getSentNotification,
   getSentNotificationDocument,
