+++ packages/pn-personafisica-webapp/src/redux/store.ts
@@ -4,6 +4,7 @@ import logger from 'redux-logger';
 import { LOG_REDUX_ACTIONS } from '../utils/constants';
 import userSlice from './auth/reducers';
 import dashboardSlice from './dashboard/reducers';
+import notificationSlice from './notification/reducers';
 
 const additionalMiddlewares = [LOG_REDUX_ACTIONS ? logger : undefined];
 
@@ -12,7 +13,8 @@ export const createStore = () =>
     reducer: {
       appState: appStateReducer,
       userState: userSlice.reducer,
-      dashboardState: dashboardSlice.reducer
+      dashboardState: dashboardSlice.reducer,
+      notificationState: notificationSlice.reducer
     },
     middleware: (getDefaultMiddleware) =>
       additionalMiddlewares.reduce(
